+++ src/Network/Network/Models/PSVirtualNetworkGatewayConnection.cs
@@ -35,6 +35,9 @@ namespace Microsoft.Azure.Commands.Network.Models
         [Ps1Xml(Target = ViewControl.Table)]
         public int RoutingWeight { get; set; }
 
+        [Ps1Xml(Target = ViewControl.Table)]
+        public int DpdTimeoutSeconds { get; set; }
+
         public string SharedKey { get; set; }
 
         [Ps1Xml(Target = ViewControl.Table)]