+++ src/Network/Network/ChangeLog.md
@@ -19,11 +19,13 @@
 --->
 
 ## Upcoming Release
+
+## Version 2.5.0
 * Updated cmdlets to enable connection on private IP for Virtual Network Gateway.
-    - New-AzVirtualNetworkGateway
-    - Set-AzVirtualNetworkGateway
-    - New-AzVirtualNetworkGatewayConnection
-    - Set-AzVirtualNetworkGatewayConnection
+    - `New-AzVirtualNetworkGateway`
+    - `Set-AzVirtualNetworkGateway`
+    - `New-AzVirtualNetworkGatewayConnection`
+    - `Set-AzVirtualNetworkGatewayConnection`
 * Updated cmdlets to enable FQDN based LocalNetworkGateways and VpnSites
     - `New-AzLocalNetworkGateway`
     - `New-AzVpnSiteLink`
@@ -32,11 +34,28 @@
         - allows setting of all the existing properties including the IPv6CircuitConnectionProperties
     - Updated `Add-AzExpressRouteCircuitConnectionConfig`
         - Added another optional parameter AddressPrefixType to specify the address family of address prefix
-* Updated cmdlets to enable connection on private IP for Virtual Network Gateway.
-    - New-AzVirtualNetworkGateway
-    - Set-AzVirtualNetworkGateway
+* Updated cmdlets to enable setting of DPD Timeout on Virtual Network Gateway Connections.
     - New-AzVirtualNetworkGatewayConnection
     - Set-AzVirtualNetworkGatewayConnection
+* Added resource type IpAllocation 
+* Added properties to Subnet
+    - Added property 'IpAllocations' as type of PSResourceId to PSIpAllocation
+* Added properties to Virtual Network
+    - Added property 'IpAllocations' as type of PSResourceId to PSIpAllocation
+* Added support for IpAllocation resource
+    - New cmdlet added:
+        - Get-AzIpAllocation
+        - New-AzIpAllocation
+        - Remove-AzIpAllocation
+        - Get-AzIpAllocation
+    - Updated `New-AzVirtualNetwork`
+        - Added another optional parameter IpAllocations to specify the IpAllocation
+    - Updated `New-AzVirtualNetworkSubnetConfig`
+        - Added another optional parameter IpAllocations to specify the IpAllocation
+    - Updated `Set-AzVirtualNetworkSubnetConfig`
+        - Added another optional parameter IpAllocations to specify the IpAllocation
+    - Updated `Add-AzVirtualNetworkSubnetConfig`
+        - Added another optional parameter IpAllocations to specify the IpAllocation
 
 ## Version 2.4.0
 * Updated cmdlets to allow cross-tenant VirtualHubVnetConnections